Update dataset_Esplanade_no_prune.py
major work on this as I moved the bulk of the ERBM to a function ERBM_run_ica which greatly cleans up the scripting. I also moved the bulk of group averaging to run_group_bulk_average which greatly cleans up the scripting. This function also does unweighted and weighted averaging. And it can be called for different data objects. Very clean.

Still lots to clean up

Also lots of flags need to be passed to the new functions