Update module_functional_connectivity.py
added image recon to the preprocessing step. So you either are working on conc in channel space, or you work on conc in image space for functional connectivity analysis

I implemented a work flow that does BP, AR and GMS filter before projection to parcels for channel space results, but does these three filters after to projection to parcels for image space results. I do this because it is too expensive to do these filters on the the >10,000 vertices.

I added plot_correlation_matrix(), plot_connectivity_circle from MNU